# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DetailedError": {
				"properties": {
					"message": {
						"type": "string"
					},
					"standardizedErrorCode": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"standardizedErrorCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServerResponse_any_": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"error": {
						"type": "string"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/DetailedError"
						},
						"type": "array"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEventDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"eventEmail": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"eventCapacity": {
						"type": "number",
						"format": "double"
					},
					"eventOwner": {
						"type": "string"
					},
					"eventOnchainId": {
						"type": "number",
						"format": "double"
					},
					"requiredApproval": {
						"type": "boolean"
					},
					"eventStartDate": {
						"type": "string"
					},
					"eventEndDate": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"eventEmail",
					"location",
					"eventCapacity",
					"eventOwner",
					"eventOnchainId",
					"requiredApproval",
					"eventStartDate",
					"eventEndDate"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "chaineventts-backend",
		"version": "1.0.0",
		"description": "Chaineventts Backend Service",
		"contact": {
			"name": "Mubarak Aminu"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/event/create": {
			"post": {
				"operationId": "CreateEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IServerResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Events Service"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEventDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}